+++ scripts/tools/serve_web.py
@@ -24,8 +24,6 @@ class H(SimpleHTTPRequestHandler):
     }
 
     def end_headers(self):
-        self.send_header("Cross-Origin-Opener-Policy", "same-origin")
-        self.send_header("Cross-Origin-Embedder-Policy", "credentialless")
         self.send_header("Cache-Control", "no-cache")
         super().end_headers()